Support for fuzzing the manifest parser
Added the load_mom binary which takes the version of the manifest and
the path to the manifest as arguments. It then parses the MoM into the
manifest struct. This way it can be used with a fuzzer like AFL or peach
to test the robustness of the manifest parsing.

Signed-off-by: John Andersen <john.s.andersen@intel.com>